+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.1 - TBD
 - Added enums `MatchAlgorithm`, `UserDefinedInstrument`
 - Added constants `UNDEF_PRICE` and `UNDEF_ORDER_SIZE`
+- Fixed `Metadata.__bytes__` method to return valid DBN
 
 ## 0.4.0 - 2023-03-24
 - Added support for Imbalance schema